New Articles request specs, plus spec refactoring

+++ app/controllers/comments_controller.rb
@@ -7,6 +7,10 @@ class CommentsController < ApplicationController
   end
 
   def create
+    unless logged_in?
+      session_notice(:danger, 'You must be logged in!', login_path) and return
+    end
+
     @article = Article.find(params[:article_id])
     @comment = @article.comments.build(comment_params)
     @comment.user = current_user
@@ -18,34 +22,50 @@ class CommentsController < ApplicationController
     end
   end
 
-  def destroy
-    session_notice(:danger, 'You must be logged in!', login_path) unless logged_in?
-
-    comment = Comment.find(params[:id])
-    comment.destroy
-
-    redirect_to comment.article
-  end
-
   def edit
     session_notice(:danger, 'You must be logged in!', login_path) unless logged_in?
 
     @comment = Comment.find(params[:id])
 
     if logged_in?
-      session_notice(:danger, 'Wrong User') unless equal_with_current_user?(comment.user)
+      session_notice(:danger, 'Wrong User') unless equal_with_current_user?(@comment.user)
     end
 
     @article = @comment.article
   end
 
   def update
-    comment = Comment.find(params[:id])
+    unless logged_in?
+      session_notice(:danger, 'You must be logged in!', login_path) and return
+    end
 
-    if comment.update(comment_params)
-      redirect_to comment.article
+    @comment = Comment.find(params[:id])
+    @article = @comment.article
+
+    if equal_with_current_user?(@comment.user)
+      if @comment.update(comment_params)
+        redirect_to @article
+      else
+        render :edit
+      end
     else
-      render :edit
+      session_notice(:danger, 'Wrong User', login_path)
+    end
+  end
+
+  def destroy
+    unless logged_in?
+      session_notice(:danger, 'You must be logged in!', login_path) and return
+    end
+
+    comment = Comment.find(params[:id])
+    article = comment.article
+
+    if equal_with_current_user?(article.user)
+      comment.destroy
+      redirect_to article
+    else
+      session_notice(:danger, 'Wrong User')
     end
   end
